% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{find_oa}
\alias{find_oa}
\title{Find a Census OA (output area) from longitude and latitude values}
\usage{
find_oa(lon, lat)
}
\arguments{
\item{lon}{numeric. A longitude (east-west) value, for example \code{-1.234}}

\item{lat}{numeric. A latitude (south-north) value, for example \code{54.567}}
}
\value{
The ONS code for a Census OA (output area) as a character string.
For example, "E00174111"
}
\description{
Find a Census OA (output area) from longitude and latitude values
}
