% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{get_full_oa_data}
\alias{get_full_oa_data}
\title{Build a table of geographical data for an OA}
\usage{
get_full_oa_data(oacd = NULL, lon = NULL, lat = NULL)
}
\arguments{
\item{oacd}{character. The ONS code for a Census OA (output area). For
example, "E00174111"}

\item{lon}{numeric. A longitude (east-west) value, for example \code{-1.234}}

\item{lat}{numeric. A latitude (south-north) value, for example \code{54.567}}
}
\value{
A tibble with 28 (expected) columns giving wider geographical data
for the selected location
}
\description{
Pass in either an OA code or a pair of longitude and latitude values
}
