% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lonlatapi-package.R
\docType{package}
\name{lonlatapi-package}
\alias{lonlatapi}
\alias{lonlatapi-package}
\title{lonlatapi: An API to return OA (Census output area) data for a given UK point location}
\description{
Finds OA (Census output area) data from longitude and latitude data.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/francisbarton/lonlatapi}
  \item Report bugs at \url{https://github.com/francisbarton/lonlatapi/issues}
}

}
\author{
\strong{Maintainer}: Fran Barton \email{fbarton@alwaysdata.net} (\href{https://orcid.org/0000-0002-5650-1176}{ORCID})

}
\keyword{internal}
